Use go mod and separate packages
