+++ computer.csv
@@ -1,4 +1,4 @@
 name,mac,ip
-Computer1,64-07-2D-BB-BB-BF,192.168.10.254:9
-Computer2,2D-F2-3D-06-17-00,192.168.10.254:9
-Computer3,FF-B3-95-62-1C-DD,192.168.10.254:9
+ExampleComputer1,64-07-2D-BB-BB-BF,192.168.10.254:9
+ExampleComputer2,2D-F2-3D-06-17-00,192.168.10.254:9
+ExampleComputer3,FF-B3-95-62-1C-DD,192.168.10.254:9
