+++ go.mod
@@ -1,0 +1,9 @@
+module github.com/daBONDi/go-rest-wol
+
+go 1.16
+
+require (
+	github.com/gocarina/gocsv v0.0.0-20201208093247-67c824bc04d4
+	github.com/gorilla/handlers v1.5.1
+	github.com/gorilla/mux v1.8.0
+)
